Fix font issue on iOS to resolve #23
Before this change, the spacing between a note letter and a note
modifier was too large on iOS. It turned out to be due to a series of
reasons: (a) the web font, "Spectral", that I chose doesn't have glyphs
defined for the unicode code points that represent musical sharps and
flats, and (b) whatever fallback font iOS uses has a crappy spacing for
those glyphs (and also is missing the glyphs for "double sharp" and
"double flat".

I fixed this by creating a new font called "FreeSerifMusic". To create
this font, I began with FreeSerif and removed a whole lot of glyphs
using FontForge. I tried to keep the glyphs I thought I might need to
use within this app, as well as basic latin glyphs just in case. Then,
by setting FreeSerifMusic as the fallback font after Spectral, things
seem to look consistent on desktop vs iOS.