Change SVG shadows to be manually created
I was observing some problems on iOS with the drop shadow filter where
DOM elements that had the filter applied were getting rendered with
lower resolution. This fix makes the process to create shadows more
manual. You have to create a separate element and blur it. That's a
little more cumbersome, but the results look better on iOS.