Upgrade some packages and remove dependency on npm
I upgraded some packages to address security issues.

I removed npm as a dependency because there were some security issues
within dependencies of npm and it didn't seem like this repo needs to
depend on npm anyway, so it seemed best to remove npm instead of upgrade
it. Honestly I'm not sure if that was the best course of action or not.
Perhaps the dependency on npm came from the initial setup with
create-react-app. Or perhaps it got in there from something else funky I
did. Anyhow, removing it seems fine. Things still work. If there are
problems later, perhaps it would be better to add it back.