Remove most of the notation functionality
Keep VexFlow installed (because I don't want to figure out how to remove
it). Feel free to remove this dependency later if needed!

Keep 'clef' property within the store (because I don't want to deal with
merge conflicts when I continue refactoring the store soon.) I'll remove
this property after I'm done refactoring the store.

I might add the notation functionality back in at some point, but for
now I think it's best to remove it because it's not a very mature
feature. I'd like to built it out more thourougly at some point and then
add it back in.