Try converting Rotator state to Mobx
I did this because I was having a problem in
Rotator:startRotating() where
pointerGrabAngle() was depending on state.center
to be set before hand. So when trying to set the initial grab angle, we
wouldn't have a center set yet, and so we wouldn't get an initial
grab angle. Moving to Mobx solved this issue to appeared to slow down
performance slightly.