Remove complex margins around Wheel
Before this change, the Wheel SVG extended underneath the Marquee and
Footer so that we could display mode labels.

After this change, the Wheel is a simple block and much easier to deal
with. It can't go underneath things now though. But oh well -- we're
giving up on that design to make things simpler.