Don't run svelte-check within snowpack
It complains about things that it shouldn't complain about. That's not
too bad during dev, because it still runs, but when I try to build, it
fails on the errors. If I can get svelte-check down to zero-errors, then
I'd like to add this back, but I suspect that will require some
bug-fixing within svelte-check itself.