Play sound when toggling base intervals
NOTE: Unfortunately, I was unable to easily get the lights to activate
when this sound plays. I'm moving on to other higher-priority features
at the moment, but I'd like to come back to this bug in the future. It seems that
it's caused by the fact that the light lives within a component that
gets re-rendered immediately upon toggle. A potential solution would be
to have a separate light placed behind the key label that doesn't
re-render when keys become active/inactive.