Fix chord polygon issue with clicks
Chord polygons were not displaying when a chord circle was pressed very
quickly after a note. This change fixes that bug, but only for click
events. Fixing it for touch events seems like it's going to be more
difficult. That might involve refactoring NoteIdSet to allow for
multiple entries of the same noteId so that we can track the fact that a
chord circle is playing a noteId *and* a key is playing that same
noteId. Might want to do the same kind of thing for lighting somehow as
well.