Improve touch rotation on Chrome
Before this change, Chrome was intermittantly treating touchstart
events as passive, which meant that trying to rotate the scale or
keyboard would result in rotation starting but not moving. (Chrome would
treat subsequent touchmove events as scrolls.) This change fixed that by
taking a somewhat drastic measure of setting `touch-action: none` as the
default everywhere. It seems to have worked, but I wouldn't be
surprised if this causes some unintended consequences. We'll see! I
tried a bunch of other things first, to no avail.

https://developers.google.com/web/updates/2017/01/scrolling-intervention